Fixes bugs #562176 and #562177 Patch #4033043 duplicated "channel key visible to outside users" and "+l accepts nonnumeric values" "make underscores valid in nicks"
the channel key is now only added to the mode display list after it is
determined that the user requesting the information is on the channel.

and +l checks to see if the requested limit has any nondigit values in
it. if so, it generates a 461, and skips to the next requested mode change.

"_" is added to the list of valid nick characters, and because i like
powers of two, the max nick length is upped to 32 characters.